requirements.txt
blurhash-python~=1.1.3
cachetools~=5.2.0
cryptography~=39.0
dj_database_url~=1.0.0
django-cache-url~=3.4.2
django-cors-headers~=3.13.0
django-debug-toolbar~=3.8.1
django-hatchway~=0.5.2
django-htmx~=1.13.0
django-oauth-toolkit~=2.2.0
django-storages[google,boto3]~=1.13.1
djangorestframework~=3.15.1
django~=4.2.0
email-validator~=1.3.0
gunicorn~=20.1.0
httpx~=0.23
markdown_it_py~=2.1.0
pillow~=9.3.0
psycopg~=3.1.8
pydantic~=1.10.2
pyld~=2.0.3
pylibmc~=1.6.3
pymemcache~=4.0.0
pytest-asyncio~=0.20.3
python-dateutil~=2.8.2
python-dotenv~=0.21.0
redis~=4.4.0
sentry-sdk~=1.11.0
sphinx~=5.3.0
urlman~=2.0.1
uvicorn[standard]~=0.27
whitenoise~=6.2.0
